Rebasing on top of main and refactoring routes
