+++ const/routes.ts
@@ -1,8 +1,15 @@
-export const CREATE_ENTITY_PATH = '/entities/new'
-export const CREATE_PROPOSAL_PATH = '/votes/new'
-export const DASHBOARD_PATH = '/dashboard'
-export const SIGN_IN_PATH = '/login'
-export const FORGOT_PASSPHRASE_PATH = '/login'
+// ADMIN
+export const ENTITY_SIGN_IN_PATH = '/login'
+export const ENTITY_FORGOT_PASSPHRASE_PATH = '/login'
+export const DASHBOARD_PATH = '/dashboard'           // TODO: entity/dashboard
+export const CREATE_ACCOUNT_PATH = '/entity/new'     // TODO: entity/new
+export const CREATE_PROPOSAL_PATH = '/votes/new' // TODO: entity/votes/new
+
+// VOTER
+export const VOTING_AUTH_FORM_PATH = '/pub/votes/auth/form' // + processId
+export const VOTING_AUTH_LINK_PATH = '/pub/votes/auth/link' // + processId + key
+export const VOTING_PATH = '/pub/votes' // + processId
+
+// GENERAL
 export const PRICING_PATH = '/pricing'
 export const ABOUT_PATH = '/about'
-export const VOTING_AUTH_FORM_PATH = '/votes/auth/form'
